bug 1618201: remove pubsub things
This removes Pub/Sub related code, configuration, dependencies, tests,
and documentation.
