--- docker/Dockerfile.pubsub
@@ -1,12 +1,0 @@
-FROM google/cloud-sdk:279.0.0-alpine
-
-# Install Java and a pubsub-emulator
-RUN apk --update add openjdk8-jre
-RUN gcloud components install pubsub-emulator beta --quiet
-
-# Copy the run script
-COPY docker/run_pubsub.sh /tmp
-
-EXPOSE 5010
-
-CMD /tmp/run_pubsub.sh
