+++ docker/config/local_dev.env
@@ -34,18 +34,6 @@ resource.metrics.markus_backends=markus.backends.logging.LoggingMetrics,markus.b
 
 # secrets.sentry.dsn=
 
-# pubsub
-# ------
-
-resource.pubsub.service_account_file=/app/docker/empty_creds.json
-resource.pubsub.project_id=local_dev_socorro
-resource.pubsub.standard_topic_name=local_dev_standard
-resource.pubsub.standard_subscription_name=local_dev_standard_sub
-resource.pubsub.priority_topic_name=local_dev_priority
-resource.pubsub.priority_subscription_name=local_dev_priority_sub
-resource.pubsub.reprocessing_topic_name=local_dev_reprocessing
-resource.pubsub.reprocessing_subscription_name=local_dev_reprocessing_sub
-
 # elasticsearch
 # -------------
 
@@ -119,8 +107,9 @@ CRASHSTORAGE_ACCESS_KEY=foo
 CRASHSTORAGE_SECRET_ACCESS_KEY=foo
 CRASHSTORAGE_BUCKET_NAME=dev_bucket
 
-CRASHPUBLISH_CLASS=antenna.ext.pubsub.crashpublish.PubSubCrashPublish
-CRASHPUBLISH_SERVICE_ACCOUNT_FILE=
-CRASHPUBLISH_PROJECT_ID=local_dev_socorro
-CRASHPUBLISH_TOPIC_NAME=local_dev_socorro_normal
-CRASHPUBLISH_SUBSCRIPTION_NAME=local_dev_socorro_sub
+CRASHPUBLISH_CLASS=antenna.ext.sqs.crashpublish.SQSCrashPublish
+CRASHPUBLISH_ENDPOINT_URL=http://localstack-sqs:4576
+CRASHPUBLISH_REGION=us-east-1
+CRASHPUBLISH_ACCESS_KEY=foo
+CRASHPUBLISH_SECRET_ACCESS_KEY=foo
+CRASHPUBLISH_QUEUE_NAME=local_dev_standard
