+++ docker/config/never_on_a_server.env
@@ -23,6 +23,3 @@ STATIC_ROOT=/tmp/crashstats-static/
 # For webapp sessions, we need to allow cookies to be sent insecurely
 # since it's using HTTP and not HTTPS.
 SESSION_COOKIE_SECURE=False
-
-# Set Pub/Sub library to use emulator
-PUBSUB_EMULATOR_HOST=pubsub:5010
