+++ docker/config/test.env
@@ -9,10 +9,14 @@ STATSD_CLIENT=django_statsd.clients.null
 # Sentry
 SENTRY_DSN=
 
-# Boto
+# boto (s3/sqs)
 resource.boto.bucket_name=crashstats-test
 destination.telemetry.bucket_name=telemetry-test
 telemetry.bucket_name=telemetry-test
+
+resource.boto.standard_queue=test_standard
+resource.boto.priority_queue=test_priority
+resource.boto.reprocessing_queue=test_reprocessing
 
 # Elasticsearch configuration
 resource.elasticsearch.elasticsearch_index=testsocorro%Y%W
@@ -20,20 +24,6 @@ resource.elasticsearch.elasticsearch_index_regex=^testsocorro[0-9]{6}$
 
 # Postgres configuration
 DATABASE_URL=postgres://postgres:aPassword@postgresql:5432/socorro_test
-
-# Pub/Sub configuration
-resource.pubsub.service_account_file=/app/docker/empty_creds.json
-resource.pubsub.standard_topic_name=test_standard
-resource.pubsub.standard_subscription_name=test_standard_sub
-resource.pubsub.priority_topic_name=test_priority
-resource.pubsub.priority_subscription_name=test_priority_sub
-resource.pubsub.reprocessing_topic_name=test_reprocessing
-resource.pubsub.reprocessing_subscription_name=test_reprocessing_sub
-
-# SQS configuration
-resource.boto.standard_queue=test_standard
-resource.boto.priority_queue=test_priority
-resource.boto.reprocessing_queue=test_reprocessing
 
 # Django tests are meant to run with debug mode disabled.
 DEBUG=False
