+++ docker/run_tests_in_docker.sh
@@ -27,7 +27,7 @@ TESTIMAGE="local/socorro_app"
 
 # Start services in background (this is idempotent)
 echo "Starting services needed by tests in the background..."
-${DC} up -d elasticsearch localstack-s3 localstack-sqs postgresql pubsub statsd
+${DC} up -d elasticsearch localstack-s3 localstack-sqs postgresql statsd
 
 # If we're running a shell, then we start up a test container with . mounted
 # to /app.
@@ -42,7 +42,8 @@ if [ "$1" == "--shell" ]; then
            --network socorro_default \
            --link socorro_elasticsearch_1 \
            --link socorro_postgresql_1 \
-           --link socorro_pubsub_1 \
+           --link socorro_localstack-sqs_1 \
+           --link socorro_localstack-s3_1 \
            --link socorro_statsd_1 \
            --env-file ./docker/config/local_dev.env \
            --env-file ./docker/config/never_on_a_server.env \
@@ -96,7 +97,8 @@ docker run \
        --network socorro_default \
        --link socorro_elasticsearch_1 \
        --link socorro_postgresql_1 \
-       --link socorro_pubsub_1 \
+       --link socorro_localstack-sqs_1 \
+       --link socorro_localstack-s3_1 \
        --link socorro_statsd_1 \
        --env-file ./docker/config/local_dev.env \
        --env-file ./docker/config/never_on_a_server.env \
