+++ requirements/default.txt
@@ -172,9 +172,6 @@ libgravatar==0.2.3 \
 dockerflow==2019.10.0 \
     --hash=sha256:44baf5c78dbf28855924c8c21c41f83147a3bd854cc0c9e1930b3b46174df6c9 \
     --hash=sha256:6c5099cb4492c43b1daa5c1421b7b39d69836c3be8bdf6b5fb2fc95b1c60696b
-google-cloud-pubsub==1.1.0 \
-    --hash=sha256:bdead87e40eba93043c6704578e946cf2794366312f936da0a2b3754234dde29 \
-    --hash=sha256:dc084349ff3fb6a58968b278e743ef9ce99ed1a72cb80576ed1d88f7fa62f808
 sentry-sdk==0.14.1 \
     --hash=sha256:b06dd27391fd11fb32f84fe054e6a64736c469514a718a99fb5ce1dff95d6b28 \
     --hash=sha256:e023da07cfbead3868e1e2ba994160517885a32dfd994fc455b118e37989479b
