+++ setup.cfg
@@ -29,10 +29,6 @@ testpaths = socorro/
 # Transform all warnings into errors
 filterwarnings =
     error
-    # grpcio kicks this up for some reason in pytest
-    ignore::ImportWarning
-    ignore:::grpcio[.*]
-    ignore:::google[.*]
     # boto3 kicks up ResourceWarning: unclosed socket, https://github.com/boto/boto3/issues/454
     ignore:unclosed:ResourceWarning
     # boto3 kicks up DeprecationWarning about collections, https://github.com/boto/botocore/issues/1885
