+++ webapp-django/pytest.ini
@@ -10,10 +10,6 @@ filterwarnings =
     # CachedStaticFilesStorage is deprecated in favor of
     # ManifestStaticFilesStorage, removal scheduled for Django 3.1
     ignore::PendingDeprecationWarning:django[.*]
-    # grpcio kicks this up for some reason in pytest
-    ignore::ImportWarning
-    ignore:::grpcio[.*]
-    ignore:::google[.*]
     # configman kicks up errors if there's configuration stuff that isn't used
     # and that breaks the webapp's configman glue
     ignore:::configman[.*]
